+++ catkit/enumeration.py
@@ -1,6 +1,6 @@
-from .gen.surface import SlabGenerator
 from .gen.surface import get_unique_indices
-from .gen import utils
+from .gen.surface import SlabGenerator
+from .gen.symmetry import get_standardized_cell
 import numpy as np
 
 
@@ -53,7 +53,7 @@ def surfaces(
     slabs : list of Gratoms objects
         Return a list of enumerated slab structures.
     """
-    standardized_bulk = utils.get_spglib_cell(bulk, tol=5e-3)
+    standardized_bulk = get_standardized_cell(bulk, tol=5e-3)
 
     if isinstance(miller_indices, int):
         miller_indices = get_unique_indices(standardized_bulk, miller_indices)
