+++ catkit/gen/utils/__init__.py
@@ -1,11 +1,9 @@
 from .connectivity import (get_voronoi_neighbors, get_cutoff_neighbors)
-from .coordinates import (trilaterate, get_unique_xy, expand_cell)
+from .coordinates import (trilaterate, get_unique_xy, expand_cell,
+                          matching_coordinates, get_unique_coordinates)
 from .geometry import (matching_sites, _get_basis_vectors,
                        _get_position, _branch_molecule)
 from .graph import (connectivity_to_edges, isomorphic_molecules)
-from .symmetry import (get_spglib_cell, get_point_group, get_symmetry,
-                       get_affine_operations, matching_coordinates,
-                       get_unique_coordinates)
 from .vectors import (get_reciprocal_vectors, plane_normal)
 from .utilities import (running_mean, to_gratoms, get_atomic_numbers,
                         get_reference_energies, parse_slice, ext_gcd,
@@ -22,10 +20,6 @@ __all__ = ['get_voronoi_neighbors',
            '_branch_molecule',
            'connectivity_to_edges',
            'isomorphic_molecules',
-           'get_spglib_cell',
-           'get_point_group',
-           'get_symmetry',
-           'get_affine_operations',
            'matching_coordinates',
            'get_unique_coordinates',
            'get_reciprocal_vectors',
